Minor updates to eviction.
Update 1: Changes the freed memory from an int to a long. Since ints are
unsigned, freed data sizes between 2 and 4 GB were returning negative
freed space, which failed a >0 check that should have passed.
Update 2: Allows eviction even if the first buffer is at a position
greater than 0. I'm not sure why that was added, but it was causing
problems when the input (synthetic) dataset size was too large to fit
into memory.

+++ seep-worker/src/main/java/uk/ac/imperial/lsds/seepworker/core/DataReferenceManager.java
@@ -249,9 +249,9 @@ public class DataReferenceManager {
 		return name;
 	}
 	
-	public int sendDatasetToDisk(int datasetId) throws IOException {
+	public long sendDatasetToDisk(int datasetId) throws IOException {
 		LOG.info("Caching Dataset to disk, id -> {}", datasetId);
-		int freedMemory = cacher.cacheToDisk(datasets.get(datasetId));
+		long freedMemory = cacher.cacheToDisk(datasets.get(datasetId));
 		LOG.info("Cached to disk, id -> {}, freedMemory -> {}", datasetId, freedMemory);
 		return freedMemory;
 	}
@@ -362,9 +362,9 @@ public class DataReferenceManager {
 		return synthetic;
 	}
 	
-	public int spillDatasetsToDisk(Integer datasetId) {
+	public long spillDatasetsToDisk(Integer datasetId) {
 		LOG.info("Worker node runs out of memory while writing to dataset: {}", datasetId);
-		int freedMemory = 0;
+		long freedMemory = 0;
 		
 		try {
 			if(rankedDatasets == null) {
